Restructure tests/ folder to match src/ folder.
