+++ tests/resources/mesh/objParser.test.ts
