+++ tests/ui/components/WebGPUWarning.test.ts
