+++ tests/ui/components/panels/ConsolePanel.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect } from 'vitest'
-import { mountWithGlobalStubs } from '../utils/mountWithGlobalStubs'
+import { mountWithGlobalStubs } from '../../utils/mountWithGlobalStubs'
 import ConsolePanel from '@/ui/components/panels/ConsolePanel.vue'
 
 describe('ConsolePanel.vue', () => {
