+++ tests/ui/components/panels/ResourcesPanel.test.ts
@@ -1,6 +1,6 @@
 import { describe, it, expect } from 'vitest'
 import ResourcesPanel from '@/ui/components/panels/ResourcesPanel.vue'
-import { mountWithGlobalStubs } from '../utils/mountWithGlobalStubs'
+import { mountWithGlobalStubs } from '../../utils/mountWithGlobalStubs'
 
 function mountResourcesPanel(customProps = {}) {
   return mountWithGlobalStubs(ResourcesPanel, {
