+++ tests/ui/utils/mountWithGlobalStubs.ts
